Fix public status of some vfs routines; add wrapper to prevent warning about NO_FILEOP_ALIAS redefinition

+++ src/filesyslib/winfiles.c
@@ -2037,7 +2037,7 @@ struct find_cursor_data {
 	const char *root;
 	const char *filemask;
 #ifdef WIN32
-   int findHandle;
+	intptr_t findHandle;
 	struct _finddata_t fileinfo;
 #else
 	DIR* handle;
